websockets-New-Audio-Transcriber/websockets-smart-consult-transcription/requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
numpy
openai-whisper
